Fix (apparently illegal?) float and short multiplication

+++ template-project/source/screen-draw-actor.c
@@ -33,7 +33,7 @@ if (!cloneindex)
     for (slice = 0; slice < screenWidth; slice++)
     {
         // calculate the position and direction of the ray
-        cameraX = 2.0f * slice / (float)screenWidth - 1; // x on the camera plane
+        cameraX = 2.0f * (float)slice / (float)screenWidth - 1; // x on the camera plane
         rayPosX = camera->pos.x; // set the begin position of the ray to the player's position
         rayPosY = camera->pos.y;
         rayDirX = camera->dir.x + camera->plane.x * cameraX; // set the direction of the ray
